mock_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-multipart
pydantic-settings
passlib[argon2]
python-jose[cryptography]
argon2-cffi
sqlalchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0
psycopg2-binary
requests
httpx
pypdf>=4.0.0
openai>=1.0.0
websockets>=12.0
azure-cognitiveservices-speech>=1.32.0
python-dotenv>=1.0.0
azure-identity
azure-mgmt-containerinstance